Backend/requirements.txt
flask
flask-cors
pymongo
geopy
requests
gunicorn
python-dotenv